Update Newest
Comment code ko xài, xóa bớt button + album list

+++ FacebookImageUpload/Form1.cs
@@ -27,12 +27,11 @@ namespace FacebookImageUpload
         public Form1()
         {
             InitializeComponent();
-            LoadingAlbumList();
+            //LoadingAlbumList();
             //LoadingAlbumList_In();
 
             cmbSelectTextType.SelectedIndex = 0;
-            cmbChangeUser.SelectedIndex = 0;
-            
+
             inbox = User.inboxUserA;
             outbox = User.inboxUserB;
             r = new Random();
@@ -48,13 +47,13 @@ namespace FacebookImageUpload
         private void openfile_Click(object sender, EventArgs e)
         {
             string filter = "Image Files(*.jpg; *.jpeg)|*.jpg; *.jpeg";
-            openfile_Click_fn(tbImagePath,filter,true);
-           
+            openfile_Click_fn(tbImagePath, filter, true);
+
         }
         private void button1_Click(object sender, EventArgs e)
         {
             string filter = "Text Files(*.txt)|*.txt";
-            openfile_Click_fn(tbMessagePath,filter);
+            openfile_Click_fn(tbMessagePath, filter);
         }
 
         private void createAlbum_Click(object sender, EventArgs e)
@@ -62,9 +61,6 @@ namespace FacebookImageUpload
             try
             {
                 FacebookAlbum album = new FacebookAlbum();
-                tbAlbumID.Text = album.createAlbum(FB_Image.AccessToken, tbAlbumDesc.Text, tbAlbumName.Text);
-                lbAlbumName.Text = album.getName(FB_Image.AccessToken);
-                pBoxAlbumCover.Image = FacebookImageUpload.Properties.Resources._default;
             }
             catch (Exception ex)
             {
@@ -74,56 +70,64 @@ namespace FacebookImageUpload
 
         private async void uploadImage_Click(object sender, EventArgs e)
         {
-            string messagePath = "";
-            if (cmbSelectTextType.SelectedIndex == cmbSelectTextType.Items.IndexOf("From File"))
+            if(lbUserIdComm.Text!="userid")
             {
-                messagePath = tbMessagePath.Text;
-            }
-            else if (cmbSelectTextType.SelectedIndex == cmbSelectTextType.Items.IndexOf("From Text"))
-            {
-                string tempMessagePath = Path.Combine(FB_Image.RelativeDirectory, "temp", "temp_message.txt");
-                File.WriteAllText(tempMessagePath, tbInputMessage.Text);
-                messagePath = tempMessagePath;
-
-            }
-
-            var progress = new Progress<string>(s => { Common.ShowProgressBar(s, pbStatus, lbStatusBar, lbDoing); });
-            //await Task.Factory.StartNew(() => GetAlbumList_1(progress, inboxAlbums, 5), TaskCreationOptions.LongRunning);
-
-            if (cbIsTested.Checked)
-            {
-                string flag="";
-                //await Task.Factory.StartNew(() => flag=SendMessageWithTestedSource(progress,tbImagePath.Text, messagePath, tbAlbumID.Text), TaskCreationOptions.LongRunning);
-                await Task.Factory.StartNew(() => flag = SendImageWithTag(progress, tbImagePath.Text, messagePath, lbUserIdComm.Text), TaskCreationOptions.LongRunning);
-                if (flag != null)
+                string messagePath = "";
+                if (cmbSelectTextType.SelectedIndex == cmbSelectTextType.Items.IndexOf("From File"))
                 {
-                    MessageBox.Show("The message has been sent");
+                    messagePath = tbMessagePath.Text;
+                }
+                else if (cmbSelectTextType.SelectedIndex == cmbSelectTextType.Items.IndexOf("From Text"))
+                {
+                    string tempMessagePath = Path.Combine(FB_Image.RelativeDirectory, "temp", "temp_message.txt");
+                    File.WriteAllText(tempMessagePath, tbInputMessage.Text);
+                    messagePath = tempMessagePath;
+
+                }
+
+                var progress = new Progress<string>(s => { Common.ShowProgressBar(s, pbStatus, lbStatusBar, lbDoing); });
+                //await Task.Factory.StartNew(() => GetAlbumList_1(progress, inboxAlbums, 5), TaskCreationOptions.LongRunning);
+
+                if (cbIsTested.Checked)
+                {
+                    string flag = "";
+                    //await Task.Factory.StartNew(() => flag=SendMessageWithTestedSource(progress,tbImagePath.Text, messagePath, tbAlbumID.Text), TaskCreationOptions.LongRunning);
+                    await Task.Factory.StartNew(() => flag = SendImageWithTag(progress, tbImagePath.Text, messagePath, lbUserIdComm.Text), TaskCreationOptions.LongRunning);
+                    if (flag != null)
+                    {
+                        MessageBox.Show("The message has been sent");
+                    }
+                    else
+                    {
+                        MessageBox.Show("This image is not ready for sending your message");
+                    }
                 }
                 else
                 {
-                    MessageBox.Show("This image is not ready for sending your message");
+                    string flag = "";
+                    await Task.Factory.StartNew(() => TestEncodeSuccessRate(progress, tbImagePath.Text, messagePath, "106782166359188", false));
+                    if (flag != null)
+                    {
+                        MessageBox.Show("The message has been sent");
+                    }
+                    else
+                    {
+                        MessageBox.Show("This image is not ready for sending your message");
+                    }
                 }
+                Common.ResetStatusTrip(pbStatus, lbStatusBar, lbDoing);
+                Common.DeleteFile(Common.listFileDelete);
             }
-            else
+            else 
             {
-                string flag = "";
-                await Task.Factory.StartNew(() => TestEncodeSuccessRate(progress, tbImagePath.Text, messagePath, tbAlbumID.Text, false));
-                if (flag != null)
-                {
-                    MessageBox.Show("The message has been sent");
-                }
-                else
-                {
-                    MessageBox.Show("This image is not ready for sending your message");
-                }
+                MessageBox.Show("Please choose Receiver from Receiver Tab", "Choose Receiver", MessageBoxButtons.OK,
+                                       MessageBoxIcon.Warning);
             }
-            Common.ResetStatusTrip(pbStatus, lbStatusBar, lbDoing);
-            Common.DeleteFile(Common.listFileDelete);      
-            
+
         }
 
 
-        private async void btngetAlbumlist_Click(object sender, EventArgs e)
+        /*private async void btngetAlbumlist_Click(object sender, EventArgs e)
         {
             btngetAlbumlist.Enabled = false;
             pbStatus.Maximum = 100;
@@ -159,10 +163,10 @@ namespace FacebookImageUpload
                 Log(ex);
             }
             btngetAlbumlist.Enabled = true;
-        }
+        }*/
 
 
-        private void ListViewalbumList_ItemActivate(object sender, EventArgs e)
+        /*private void ListViewalbumList_ItemActivate(object sender, EventArgs e)
         {
             DialogResult dlg = MessageBox.Show("Do you want to choose this Album for upload?", "Choose Album", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
             if (dlg == DialogResult.Yes)
@@ -172,7 +176,7 @@ namespace FacebookImageUpload
                 lbAlbumName.Text = item.Text.ToString();
                 tbAlbumID.Text = item.Name.ToString();
             }
-        }
+        }*/
 
 
         private void cmbSelectTextType_SelectedIndexChanged(object sender, EventArgs e)
@@ -182,7 +186,7 @@ namespace FacebookImageUpload
 
         private void tbMessage_KeyPress(object sender, KeyPressEventArgs e)
         {
-            if(isFromFile)
+            if (isFromFile)
                 e.Handled = true;
         }
 
@@ -191,7 +195,7 @@ namespace FacebookImageUpload
          * 
          *  */
 
-        private async void btnGetAlbumInbox_Click(object sender, EventArgs e)
+       /* private async void btnGetAlbumInbox_Click(object sender, EventArgs e)
         {
             btnGetAlbumInbox.Enabled = false;
             pbStatus.Maximum = 100;
@@ -228,7 +232,7 @@ namespace FacebookImageUpload
             }
             btnGetAlbumInbox.Enabled = true;
 
-        }
+        }*/
 
         private string GetInboxText(string name, int count)
         {
@@ -244,9 +248,9 @@ namespace FacebookImageUpload
             return text;
         }
 
-        private void btnChangeUser_Click(object sender, EventArgs e)
+        /*private void btnChangeUser_Click(object sender, EventArgs e)
         {
-         
+
             if (cmbChangeUser.SelectedItem.Equals("User A"))
             {
                 inbox = User.inboxUserA;
@@ -258,9 +262,9 @@ namespace FacebookImageUpload
                 outbox = User.inboxUserA;
             }
 
-        }
+        }*/
 
-        private async void btnGetMessage_Click(object sender, EventArgs e)
+        /*private async void btnGetMessage_Click(object sender, EventArgs e)
         {
 
             var a = ListViewalbumList_In.SelectedItems;
@@ -271,11 +275,11 @@ namespace FacebookImageUpload
                 {
                     string albumID = i.Name;
                     string text = i.Text;
-                    if( text.IndexOf('-')>0)
+                    if (text.IndexOf('-') > 0)
                         text = text.Substring(0, text.IndexOf('-'));
-                    List<FB_Message> listMessage=null;
+                    List<FB_Message> listMessage = null;
                     var progress = new Progress<string>(s => { });
-                    await Task.Factory.StartNew(()=>{listMessage= GetNewMessageFromAlbum(progress,albumID);},TaskCreationOptions.LongRunning);
+                    await Task.Factory.StartNew(() => { listMessage = GetNewMessageFromAlbum(progress, albumID); }, TaskCreationOptions.LongRunning);
                     if (listMessage == null)
                     {
                         tbInbox.AppendText("No ouput" + Environment.NewLine);
@@ -301,7 +305,7 @@ namespace FacebookImageUpload
 
                 MessageBox.Show("get message is finished");
             }
-        }
+        }*/
 
 
         private async void btnImageSearch_Click(object sender, EventArgs e)
@@ -310,10 +314,11 @@ namespace FacebookImageUpload
             try
             {
                 var progress = new Progress<string>
-                    (s => { 
+                    (s =>
+                    {
                         progressString = s;
                         tbGoogleLink.AppendText(progressString + Environment.NewLine);
-                });
+                    });
                 await Task.Factory.StartNew(() => btnImageSearch_Click_fn(progress), TaskCreationOptions.LongRunning);
                 MessageBox.Show("Searching is finish");
 
@@ -322,7 +327,7 @@ namespace FacebookImageUpload
             {
                 Log(ex);
             }
-            
+
         }
 
         private void tbInputMessage_TextChanged(object sender, EventArgs e)
@@ -372,7 +377,7 @@ namespace FacebookImageUpload
                 {
                     List<FB_Message> listMessage = null;
                     var progress = new Progress<string>(s => { });
-                    await Task.Factory.StartNew(() => { listMessage = GetNewMessageFromImage(progress,i.Text); }, TaskCreationOptions.LongRunning);
+                    await Task.Factory.StartNew(() => { listMessage = GetNewMessageFromImage(progress, i.Text); }, TaskCreationOptions.LongRunning);
                     if (listMessage == null)
                     {
                         tbInbox.AppendText("No ouput" + Environment.NewLine);
@@ -397,6 +402,8 @@ namespace FacebookImageUpload
 
                 MessageBox.Show("get message is finished");
             }
-        }  
+        }
+
     }
+
 }
